refactor to put chartmuseum's relate function to separated package
