+++ plugin.go
@@ -4,6 +4,7 @@ import (
 	"io/ioutil"
 	"log"
 
+	"github.com/honestbee/drone-chartmuseum/pkg/chartmuseumclient"
 	"github.com/honestbee/drone-chartmuseum/pkg/util"
 )
 
@@ -28,12 +29,12 @@ type (
 func (p *Plugin) defaultExec(files []string) {
 	var resultList []string
 	for _, file := range files {
-		chart, err := util.SaveChartToPackage(file, p.Config.SaveDir)
+		chart, err := chartmuseumclient.SaveChartToPackage(file, p.Config.SaveDir)
 		if err == nil {
 			resultList = append(resultList, chart)
 		}
 	}
-	util.UploadToServer(resultList, p.Config.RepoURL)
+	chartmuseumclient.UploadToServer(resultList, p.Config.RepoURL)
 }
 
 func (p *Plugin) exec() error {
@@ -41,7 +42,8 @@ func (p *Plugin) exec() error {
 	if p.Config.ChartPath != "" {
 		files = []string{p.Config.ChartPath}
 	} else if p.Config.PreviousCommitID != "" && p.Config.CurrentCommitID != "" {
-		files = util.GetParentFolders(util.FilterExtFiles(util.GetDiffFiles(p.Config.ChartPath, p.Config.PreviousCommitID, p.Config.CurrentCommitID)))
+		diffFiles := chartmuseumclient.GetDiffFiles(p.Config.ChartPath, p.Config.PreviousCommitID, p.Config.CurrentCommitID)
+		files = util.GetParentFolders(util.FilterExtFiles(diffFiles))
 	} else {
 		dirs, err := ioutil.ReadDir(p.Config.ChartDir)
 		if err != nil {
